requirements.txt
Django==5.2.*
psycopg[binary]
djangorestframework
gunicorn

